+++ frontend/src/App.jsx
@@ -5,14 +5,13 @@ import {
   Route,
   Navigate
 } from 'react-router-dom';
-import Login from './Login';
-import MainLayout from './MainLayout';
-import ApplicationsPage from './ApplicationsPage';
+import Login from './pages/Login';
+import HomePage from './pages/HomePage';
+import Services from './pages/Services';
 import './App.css';
 import VoiceToTest from './pages/voicetotest';
 import TTS from './pages/TTS';
-import Heygen from './pages/heygen';
-import SettingsPage from './SettingsPage';
+import SettingsPage from './pages/SettingsPage';
 
 function App() {
   // For now, use localStorage for auth state (simple demo)
@@ -25,40 +24,50 @@ function App() {
   return (
     <Router>
       <Routes>
+
         <Route 
           path="/login" element={<Login />} />
+
         <Route
+          // path="/applications"
           path="/applications"
           element={
             role === 'user'
-              ? <ApplicationsPage />
+              ? <HomePage />
               : <Navigate to={role === 'admin' ? '/settings' : '/login'} />}
         />
+
         <Route
-          path="/pingfederate/*"
-          element={isAuthenticated ? <MainLayout /> : <Navigate to="/login" />}
+          // path="/pingfederate/*"
+          path = "/services/*"
+          element={isAuthenticated ? <Services /> : <Navigate to="/login" />}
         />
+
         <Route
           path="/voicetotest"
           element={isAuthenticated ? <VoiceToTest /> : <Navigate to="/login" />}
         />
+
         <Route
           path="/tts"
           element={isAuthenticated ? <TTS /> : <Navigate to="/login" />}
         />
-        <Route
-          path="/videoavatar"
-          element={isAuthenticated ? <Heygen /> : <Navigate to="/login" />}
-        />
+
+        {/* <Route
+          path="/aiavatar"
+          element={isAuthenticated ? <AvatarPage /> : <Navigate to="/login" />}
+        /> */}
+
         <Route
           path="/*"
           element={<Navigate to={isAuthenticated ? "/applications" : "/login"} />}
         />
+        
         <Route
         path="/settings"
         element={
           role === 'admin'
-            ? <MainLayout><SettingsPage /></MainLayout>
+            ? <SettingsPage />
             : <Navigate to={role === 'user' ? "/applications" : "/login"} />
         }
         />
