--- frontend/src/MainLayout.css
@@ -1,286 +1,0 @@
-html, body, #root {
-  width: 100vw;
-  height: 100vh;
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  background: #e3eaf0;
-  overflow-x: hidden;
-}
-
-.main-layout {
-  
-  display: flex;
-  height: 100vh;
-  background: #e3eaf0;
-  overflow-x: hidden;
-}
-
-.main-content {
-  position: absolute;
-  top: 64px; /* Height of the top bar */
-  left: 235px; /* Width of the sidebar */
-  width: calc(100vw - 260px);
-  height: calc(100vh - 64px);
-  padding-left: 3%;
-  display: flex;
-  flex-direction: column;
-  background: #e3eaf0;
-  box-sizing: border-box;
-  overflow-y: auto;
-  overflow-x: hidden;
-}
-
-.main-content-chat {
-  background: #e3eaf0;
-  width: 100%;
-  height: 100%;
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-  
-}
-
-.main-header {
-  width: 100%;
-  background: #fff !important;
-  padding: 18px 32px 18px 0;
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  box-shadow: 0 2px 8px rgba(0,0,0,0.03);
-  min-height: 100px;
-  position: relative !important;
-  z-index: 9998 !important;
-}
-.user-info {
-  display: flex;
-  align-items: center;
-  position: relative;
-  z-index: 3;
-}
-.user-avatar {
-  width: 48px;
-  height: 48px;
-  border-radius: 50%;
-  margin-right: 16px;
-}
-.user-avatar.user-avatar-circle {
-  width: 48px !important;
-  height: 48px !important;
-  border-radius: 50% !important;
-  background: #6ea3e7 !important;
-  color: #fff !important;
-  display: flex !important;
-  align-items: center !important;
-  justify-content: center !important;
-  font-size: 1.5em !important;
-  font-weight: bold !important;
-  margin-right: 8px !important;
-  user-select: none !important;
-  border: 3px solid #fff !important;
-  z-index: 9999 !important;
-  position: relative !important;
-  box-shadow: 0 2px 8px rgba(0,0,0,0.08) !important;
-}
-.user-avatar.user-avatar-circle:hover {
-  background: #4b7fc9 !important;
-  border-color: #4b7fc9 !important;
-}
-.greeting {
-  font-size: 0.95em;
-  color: #ccc3c3;
-}
-.username {
-  font-weight: bold;
-  font-size: 1.1em;
-}
-
-.page-title-container {
-  width: 100%;
-  margin-left: 0;
-  padding-left: 32px;
-  margin-top: 32px;
-  margin-bottom: 16px;
-  display: flex;
-  align-items: flex-start;
-}
-.page-title {
-  font-size: 2.5em;
-  font-weight: 700;
-  color: #23242a;
-  margin: 0;
-}
-.applications-section {
-  flex: 1;
-  padding: 0 0 32px 14px;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  justify-content: flex-start;
-  width: 100%;
-  min-height: fit-content;
-  box-sizing: border-box;
-  overflow-y: auto;
-}
-.applications-title {
-  font-size: 2.5em;
-  font-weight: 700;
-  color: #23242a;
-  margin-bottom: 8px;
-}
-.applications-subtitle {
-  font-size: 1.25em;
-  color: #23242a;
-  font-weight: 500;
-  margin-bottom: 32px;
-}
-.cards-row {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 40px;
-  width: calc(100% - 32px);
-  padding: 0 16px 32px 0;
-  margin-bottom: 32px;
-  box-sizing: border-box;
-}
-
-@media (min-width: 2000px) {
-  .cards-row {
-    grid-template-columns: repeat(4, 1fr);
-  }
-}
-.app-card {
-  background: #fff;
-  height: auto;
-  width: 280px;
-  max-width: 320px;
-  margin: 0 auto;
-  border-radius: 16px;
-  box-shadow: 0 2px 8px rgba(0,0,0,0.08);
-  padding: 32px 24px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  transition: box-shadow 0.2s, transform 0.2s;
-  /* color: black; Text Colour */
-}
-.app-card:hover {
-  box-shadow: 0 4px 16px rgba(0,0,0,0.13);
-  transform: translateY(-2px);
-}
-.card-title {
-  font-size: 1.2em;
-  font-weight: 600;
-  margin-bottom: 18px;
-}
-.card-img {
-  width: 140px;
-  height: 140px;
-  object-fit: contain;
-}
-
-.main-layout.full-screen {
-  width: 100vw;
-  height: 100vh;
-  min-height: 100vh;
-  min-width: 100vw;
-  background: #e3eaf0;
-  display: flex;
-}
-
-.top-header {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 64px;
-  background: #946cba;
-  color: #fff;
-  display: flex;
-  align-items: center;
-  justify-content: flex-start;
-  z-index: 1000;
-  box-shadow: 0 2px 8px rgba(0,0,0,0.08);
-}
-.header-logo {
-  margin-left: 24px;
-  margin-right: 24px;
-  display: flex;
-  align-items: center;
-  height: 64px;
-}
-.header-logo img {
-  height: 44px;
-  width: auto;
-  object-fit: contain;
-}
-.header-app-name {
-  flex: 1;
-  text-align: center;
-  font-size: 1.35em;
-  font-weight: 600;
-  letter-spacing: 0.5px;
-}
-.header-user-info {
-  position: absolute;
-  right: 32px;
-  top: 0;
-  height: 64px;
-  display: flex;
-  align-items: center;
-  z-index: 1100;
-}
-.header-user-text {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  margin-left: 8px;
-}
-
-.logout-dropdown {
-  position: absolute;
-  top: 56px;
-  right: 0;
-  background: #fff;
-  color: #333;
-  border: 1px solid #eee;
-  border-radius: 8px;
-  box-shadow: 0 2px 8px rgba(0,0,0,0.08);
-  padding: 10px 24px;
-  cursor: pointer;
-  z-index: 10000;
-  font-size: 1em;
-  transition: background 0.2s;
-}
-.logout-dropdown:hover {
-  background: #f5f5f5;
-}
-
-.login-container {
-  position: fixed;
-  top: 0; left: 0;
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: #fff;
-}
-
-.back-arrow-btn {
-  background: none;
-  border: none;
-  color: #23242a;
-  font-size: 2em;
-  margin-right: 12px;
-  cursor: pointer;
-  padding: 0 8px 0 0;
-  transition: color 0.2s;
-  outline: none;
-  vertical-align: middle;
-}
-.back-arrow-btn:hover {
-  color: #6ea3e7;
-}
