+++ frontend/src/components/sidebarLayout.css
@@ -1,6 +1,6 @@
 
 .sidebar {
-  width: 260px;
+  width: 220px;
   background: linear-gradient(180deg, #6ea3e7 0%, #946cba 100%);
   color: #fff;
   display: flex;
