+++ frontend/src/pages/ChatPage.css
@@ -1,0 +1,337 @@
+.chat-container {
+  position: absolute;
+  top: 64px;
+  left: 260px;
+  height: calc(100vh - 64px);
+  width: calc(100vw - 280px);
+  max-width: calc(100vw - 280px);
+  display: flex;
+  background-color: white;
+  overflow: hidden;
+  box-sizing: border-box;
+}
+
+/* Fills the sidebar space + stretches */
+.chat-left {
+  width: 40%;
+  min-width: 0;
+  height: 100%;
+  background: linear-gradient(to bottom, #dbeafe, #eef2ff);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-between;
+  padding: 40px 20px;
+  border-right: 1px solid #ddd;
+  box-sizing: border-box;
+  overflow: hidden;
+}
+
+/* Avatar takes full vertical space available */
+.avatar-full-rectangle {
+  width: 100%;
+  height: 100%;
+  max-height: 400px;
+  background-color: #cfcfcf;
+  border-radius: 12px;
+  margin-bottom: 30px;
+}
+
+/* Circular icon buttons */
+.chat-icon-buttons {
+  display: flex;
+  gap: 16px;
+  justify-content: center;
+  margin-bottom: 20px;
+}
+
+.round-button {
+  padding: 0;
+  width: 48px;
+  height: 48px;
+  border-radius: 50%;
+  border: none;
+  background-color: #6366f1;
+  color: white;
+  font-size: 1.2rem;
+  cursor: pointer;
+  box-shadow: 0 2px 6px rgba(0,0,0,0.1);
+  transition: background 0.2s;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.round-button:hover {
+  background-color: #4f46e5;
+}
+
+/* RIGHT PANEL */
+.chat-right {
+  width: 60%;
+  min-width: 0;
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  overflow: hidden;
+  box-sizing: border-box;
+}
+
+.chat-header {
+  padding: 20px 30px 10px;
+  border-bottom: 1px solid #eee;
+}
+
+.chat-header h2 {
+  margin: 0;
+  font-size: 1.3rem;
+  color: #111827;
+}
+
+.chat-header p {
+  margin: 5px 0 0;
+  font-size: 0.95rem;
+  color: #6b7280;
+}
+
+.chat-messages {
+  flex-grow: 1;
+  overflow-y: auto;
+  overflow-x: hidden;
+  padding: 20px 30px;
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  min-height: 0;
+  max-width: 100%;
+  box-sizing: border-box;
+}
+
+.chat-bubble {
+  max-width: 70%;
+  padding: 12px 16px;
+  font-size: 0.95rem;
+  border-radius: 18px;
+  word-break: break-word;
+}
+
+.chat-bubble.user {
+  align-self: flex-end;
+  background-color: #946cba;
+  color: white;
+  border-bottom-right-radius: 4px;
+}
+
+.chat-bubble.bot {
+  align-self: flex-start;
+  background-color: #f3f4f6;
+  color: #111827;
+  border-bottom-left-radius: 4px;
+}
+
+/* Loading and Typing Indicator Styles */
+.chat-bubble.loading {
+  opacity: 0.7;
+  font-style: italic;
+}
+
+/* Typing indicator animation */
+.typing-indicator {
+  display: inline-flex;
+  align-items: center;
+  margin-right: 0.5rem;
+}
+
+.typing-indicator span {
+  height: 6px;
+  width: 6px;
+  background: #6366f1;
+  border-radius: 50%;
+  display: inline-block;
+  margin-right: 2px;
+  animation: typing 1.4s infinite ease-in-out;
+}
+
+.typing-indicator span:nth-child(1) {
+  animation-delay: 0s;
+}
+
+.typing-indicator span:nth-child(2) {
+  animation-delay: 0.2s;
+}
+
+.typing-indicator span:nth-child(3) {
+  animation-delay: 0.4s;
+  margin-right: 0;
+}
+
+@keyframes typing {
+  0%, 60%, 100% {
+    transform: translateY(0);
+    opacity: 0.5;
+  }
+  30% {
+    transform: translateY(-10px);
+    opacity: 1;
+  }
+}
+
+/* Markdown styling within bot chat bubbles */
+.chat-bubble.bot h1, .chat-bubble.bot h2, .chat-bubble.bot h3 {
+  margin: 0.5rem 0;
+  color: #111827;
+}
+
+.chat-bubble.bot h1 { font-size: 1.25rem; }
+.chat-bubble.bot h2 { font-size: 1.1rem; }
+.chat-bubble.bot h3 { font-size: 1rem; }
+
+.chat-bubble.bot p {
+  margin: 0.5rem 0;
+  line-height: 1.6;
+}
+
+.chat-bubble.bot ul, .chat-bubble.bot ol {
+  margin: 0.5rem 0;
+  padding-left: 1.5rem;
+}
+
+.chat-bubble.bot li {
+  margin: 0.25rem 0;
+  line-height: 1.5;
+}
+
+.chat-bubble.bot strong {
+  font-weight: 600;
+  color: #111827;
+}
+
+.chat-bubble.bot code {
+  background: rgba(0, 0, 0, 0.1);
+  padding: 0.125rem 0.25rem;
+  border-radius: 3px;
+  font-family: 'Courier New', monospace;
+  font-size: 0.9rem;
+}
+
+.chat-bubble.bot pre {
+  background: rgba(0, 0, 0, 0.05);
+  padding: 0.75rem;
+  border-radius: 6px;
+  overflow-x: auto;
+  margin: 0.5rem 0;
+}
+
+.chat-bubble.bot pre code {
+  background: none;
+  padding: 0;
+}
+
+.chat-bubble.bot blockquote {
+  border-left: 3px solid #6366f1;
+  padding-left: 1rem;
+  margin: 0.5rem 0;
+  font-style: italic;
+  color: #4b5563;
+}
+
+/* Disabled input and button states */
+.chat-input-area input:disabled {
+  background: #f9fafb;
+  color: #6b7280;
+  cursor: not-allowed;
+}
+
+.chat-input-area button:disabled {
+  background: #d1d5db;
+  color: #6b7280;
+  cursor: not-allowed;
+}
+
+.chat-input-area button:disabled:hover {
+  background: #d1d5db;
+}
+
+.chat-input-area {
+  display: flex;
+  padding: 16px 30px;
+  border-top: 1px solid #eee;
+  background-color: #fafafa;
+}
+
+.chat-input-area input {
+  flex-grow: 1;
+  padding: 10px 15px;
+  font-size: 1rem;
+  border-radius: 20px;
+  border: 1px solid #ccc;
+  margin-right: 10px;
+  background-color: #fafafa;
+  color: black;
+}
+
+.chat-input-area button {
+  padding: 10px 18px;
+  background-color: #6366f1;
+  color: white;
+  font-size: 1rem;
+  border-radius: 20px;
+  border: none;
+  cursor: pointer;
+}
+
+.chat-input-area button:hover {
+  background-color: #4f46e5;
+}
+
+/* Scrollbar Styling */
+.chat-messages::-webkit-scrollbar {
+  width: 6px;
+}
+
+.chat-messages::-webkit-scrollbar-track {
+  background: #f1f1f1;
+  border-radius: 3px;
+}
+
+.chat-messages::-webkit-scrollbar-thumb {
+  background: #6366f1;
+  border-radius: 3px;
+}
+
+.chat-messages::-webkit-scrollbar-thumb:hover {
+  background: #4f46e5;
+}
+
+.chat-messages::-webkit-scrollbar {
+  width: 6px;
+}
+.chat-messages::-webkit-scrollbar-thumb {
+  background-color: rgba(0, 0, 0, 0.1);
+  border-radius: 3px;
+}
+.chat-back-header {
+  display: flex;
+  align-items: center;
+  padding: 12px 30px 0;
+  gap: 10px;
+}
+
+.chat-back-button {
+  background: none;
+  border: none;
+  font-size: 1.2rem;
+  color: black;
+  cursor: pointer;
+  padding: 4px 8px;
+}
+
+.chat-back-button:hover {
+  color: #4b5563; /* A slightly lighter black */
+}
+
+.chat-back-label {
+  font-size: 1rem;
+  color: #111827;
+  font-weight: 500;
+}
