+++ frontend/src/pages/Login.css
@@ -1,0 +1,15 @@
+.login-container.plain-bg {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background: #fff;
+  margin: 0;
+  padding: 0;
+  overflow: hidden;
+  color: black;
+}
