--- frontend/src/pages/heygen.jsx
@@ -1,11 +1,0 @@
-function heygen() {
-    return (
-        <div style={{ display: "flex", flexDirection: "column", rowGap: "1rem", color: 'black' }}>
-            <button onClick={() => window.location.href = '/applications'} style={{ alignSelf: 'flex-start' }}>
-                Back to Applications
-            </button>
-            heygen
-        </div>
-    )
-}
-export default heygen;
